% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Minimal annotator (toy)}
\usage{
annotate_variants(df)
}
\arguments{
\item{df}{output of read_vcf()}
}
\value{
df with dummy gene/impact columns (for vignette/tests)
}
\description{
Minimal annotator (toy)
}
