% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_gene.R
\name{assoc_gene}
\alias{assoc_gene}
\title{Simple gene-level logistic regression (toy)}
\usage{
assoc_gene(var_df, pheno, gene)
}
\arguments{
\item{var_df}{annotated variant table}

\item{pheno}{data.frame(SAMPLE, PHENO\link{, AGE, SEX})}

\item{gene}{character}
}
\value{
data.frame(term, beta, se, p)
}
\description{
Simple gene-level logistic regression (toy)
}
