% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_gwas.R
\name{assoc_gwas}
\alias{assoc_gwas}
\title{Toy GWAS scan over positions (presence/absence)}
\usage{
assoc_gwas(var_df, pheno)
}
\arguments{
\item{var_df}{annotated variant table}

\item{pheno}{data.frame(SAMPLE, PHENO)}
}
\value{
data.frame(CHR, POS, P, BETA)
}
\description{
Toy GWAS scan over positions (presence/absence)
}
