% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_single.R
\name{assoc_single}
\alias{assoc_single}
\title{Association for one variant and one phenotype}
\usage{
assoc_single(
  df,
  outcome,
  dosage = "Dosage",
  covars = character(),
  model = c("logistic", "linear", "firth")
)
}
\arguments{
\item{df}{Analysis data containing outcome, dosage, optional covars.}

\item{outcome}{Outcome column name.}

\item{dosage}{Dosage column name (0/1/2).}

\item{covars}{Optional covariates.}

\item{model}{"logistic","linear","firth".}
}
\value{
list(tidy=data.frame, model=fit)
}
\description{
Fit Linear/Logistic or Firth Logistic (if \pkg{logistf} installed).
}
\details{
assoc_single
}
\examples{
\dontrun{
dat <- merge(pheno_df, subset(v$genotypes, CHROM=="1" & POS==1234)[,c("Sample","Dosage")],
             by.x="sample_id", by.y="Sample")
assoc_single(dat, outcome="trait", dosage="Dosage", covars=c("age","sex"), model="logistic")
}
}
