% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{bind_rows_list}
\alias{bind_rows_list}
\title{Bind list of data frames without rownames}
\usage{
bind_rows_list(lst)
}
\description{
Internal utility to rbind a list of data.frames and
remove row names for consistent formatting.
}
\keyword{internal}
