% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{firth_fit_safe}
\alias{firth_fit_safe}
\title{Safe Firth logistic regression}
\usage{
firth_fit_safe(formula, data, maxit)
}
\arguments{
\item{formula}{Model formula.}

\item{data}{Data frame.}

\item{maxit}{Maximum iterations.}
}
\description{
Suppresses warnings and catches errors during
Firth logistic regression fitting using \code{logistf::logistf()}.
}
\keyword{internal}
