% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read a VCF (plain or bgz) into a tidy table}
\usage{
read_vcf(path)
}
\arguments{
\item{path}{Path to .vcf or .vcf.gz}
}
\value{
data.frame with CHROM, POS, REF, ALT, SAMPLE, GT, DP, AD
}
\description{
Read a VCF (plain or bgz) into a tidy table
}
\examples{
v <- read_vcf(system.file("extdata","toy.vcf", package="VcfAssociation"))
head(v)
}
